example cleanup: dropped js code bridge and now use mrhyde 0.5.1
mrhyde 0.5.1 has Array.filter support, needed by rickshaw.
There is still a lingering bug where the series has to be
defined by a js array instead of a cljs vector - I think a
'this' is not being set right somewhers. Anyway, with this
workaround the bridge code in index.html is no longer
necessary, so the code is almost presentable now.